fix bug Given_OpenApiDocument_Then_It_Should_Return_ComponentSchemaProperty propertyFormat decimal -> double

+++ test-integration/Microsoft.Azure.WebJobs.Extensions.OpenApi.Document.Tests/Post_AppicationJson_Decimal_Tests.cs
@@ -98,7 +98,7 @@ namespace Microsoft.Azure.WebJobs.Extensions.OpenApi.Document.Tests
         }
 
         [DataTestMethod]
-        [DataRow("decimalTypeObjectModel", "decimalValue", "number", "decimal")]
+        [DataRow("decimalTypeObjectModel", "decimalValue", "number", "double")]
         public void Given_OpenApiDocument_Then_It_Should_Return_ComponentSchemaProperty(string @ref, string propertyName, string propertyType, string propertyFormat)
         {
             var properties = this._doc["components"]["schemas"][@ref]["properties"];
